Write Create and Update Person in Application layer and debug some minor bugs

+++ PersonalFinanceApp.Application/Features/Expenditures/Commands/CreateExpenditure/CreateExpenditureCommand.cs
@@ -11,7 +11,7 @@ namespace PersonalFinanceApp.Application.Features.Expenditures.Commands.CreateEx
 public class CreateExpenditureCommand : IRequest<Guid>, IExpenditureRequest
 {
     public DateOnly DocumentDate { get; set; }
-    public byte CurrencyId { get; set; }
+    public int CurrencyId { get; set; }
 
 
     // What was it spent on, and how much for each item. Every line becomes a debit entry.
