+++ PersonalFinanceApp.Application/Features/Languages/Commands/CreateLanguage/CreateLanguageCommand.cs
@@ -6,7 +6,7 @@ using MediatR;
 
 namespace PersonalFinanceApp.Application.Features.Languages.Commands.CreateLanguage;
 
-public class CreateLanguageCommand : IRequest<byte>
+public class CreateLanguageCommand : IRequest<int>
 {
     public string Code { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
